% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest}
\alias{backtest}
\title{Backtest a Set of Daily Weights}
\usage{
backtest(data, rebal_group_by = "date", rebal_ith = 1,
  prices = tidalprice::dailyclose)
}
\arguments{
\item{data}{a data.frame containing the columns: symbol, date, w}

\item{rebal_group_by}{Optional Character indicating by group to apply to ith_date.
Acceptable values are: date, week, month, qtr.  Default is date
which performs a daily rebalance.}

\item{rebal_ith}{Optional Integer to determine which observation of each by group to return.
For Example ith_date = 2, week will return the second day of
each week.  Also, ith_date = 99, month will return last day of
each month.  Default is 1.}

\item{prices}{a data.frame containing the columns: symbol, date, price.
Default is tidalprices::dailyclose.}
}
\description{
Backtest function that returns a daily NAV dataset
}
\examples{
\dontrun{
# Create a data.frame of weights
df_weights <- dplyr::filter(dailyclose, symbol \%in\% c('XLB', 'XLE', 'XLF'), date <= as.Date('1998-12-31'))
df_weights$w <- 0.32
df_weights$price <- NULL

# Trade the weights
df_nav <- backtest(df_weights)
}
}
