% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_execute.R
\name{backtest_execute}
\alias{backtest_execute}
\title{Unexported Subfunction(s) of backtest}
\usage{
backtest_execute(list_data)
}
\arguments{
\item{list_data}{list returned from backtest_initialize()}
}
\description{
Takes the list returned from backtest_initialize()
and executes the daily trades.
}
