% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_initialize.R
\name{backtest_initialize}
\alias{backtest_initialize}
\title{Unexported Subfunction(s) of backtest}
\usage{
backtest_initialize(df_prices, df_weights, use_cash_long, use_cash_short)
}
\arguments{
\item{df_prices}{a data.frame containing the columns: symbol, date, price}

\item{df_weights}{a data.frame containing the columns: symbol, date, w}

\item{use_cash_long}{if TRUE use _CASH_LONG_ price within df_prices, else use 1 for all dates}

\item{use_cash_short}{if TRUE use _CASH_SHORT_ price within df_prices, else use 1 for all dates}
}
\description{
Takes a set of daily prices and daily weights
and returns a list of initialized data ready to be traded.
}
