% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedfunds_functions.R
\name{convert_g_price}
\alias{convert_g_price}
\title{Unexported function(s) used to munge fedfunds and related data}
\usage{
convert_g_price(df)
}
\arguments{
\item{df}{data.frame containing columns: symbol, date, g (daily growth, e.g. 0.01 = 1 percent).}
}
\description{
Functions used to change fedfunds growth data into price style data.
}
