% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedfunds_functions.R
\name{convert_rate_g}
\alias{convert_rate_g}
\title{Unexported function(s) to munge fedfunds and related data}
\usage{
convert_rate_g(df)
}
\arguments{
\item{df}{data.frame containing columns: symbol, date, rate (annualized, e.g. 2.25).}
}
\description{
Changes fedfunds style data into price style data.
}
