% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedfunds_functions.R
\name{create_cash_assets}
\alias{create_cash_assets}
\title{Unexported function(s) to munge fedfunds and related data}
\usage{
create_cash_assets(df, add_rate_long = -0.5, add_rate_short = 0.25)
}
\arguments{
\item{df}{data.frame containing columns: symbol, date, rate (annualized, e.g. 2.25).}

\item{add_rate_long}{Annualized rate to add to input rate to create annualized rate
paid on cash balances(e.g. -0.50 = -25 basis points).}

\item{add_rate_short}{Annualized rate to add to input rate to create annualized rate
charge on margin balances(e.g. 0.25 = +25 basis points).}
}
\description{
Takes a benchmark interest rate like fedfunds and return broker style
cash assets with symbols "_CASH_LONG_" and "_CASH_SHORT_" in the format
that trade_weights() requires.
}
