% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyclosecash.R
\docType{data}
\name{dailyclosecash}
\alias{dailyclosecash}
\title{Daily Close Prices for Cash Assets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 730 rows and 5 columns.}
\usage{
dailyclosecash
}
\description{
Daily close prices for long and short cash assets from 2017-10-23 through 2018-10-22.
}
\keyword{datasets}
