% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailycloseetf.R
\docType{data}
\name{dailycloseetf}
\alias{dailycloseetf}
\title{Daily Close Prices for XLB, XLE, and XLF}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 756 rows and 3 columns.}
\usage{
dailycloseetf
}
\description{
Daily close prices for XLB, XLE, and XLF from 2017-10-23 through 2018-10-22.
}
\keyword{datasets}
