% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedfunds.R
\docType{data}
\name{fedfunds}
\alias{fedfunds}
\title{Daily Federal Funds Target Range - Upper Limit}
\format{An object of class \code{data.frame} with 365 rows and 3 columns.}
\source{
Raw data was downloaded from FRED via quantmod::getSymbols()
and converted to R data.frame via the code shown in the examples.
}
\usage{
fedfunds
}
\description{
Daily Federal Funds Target Range - Upper Limit
from 2017-10-23 through 2018-10-22.
}
\examples{

print(fedfunds)

# Code used to generate the fedfunds data.frame from quantmod::getSymbols()
\dontrun{
library(quantmod)
getSymbols("DFEDTARU", src = "FRED")
DFEDTARU <- DFEDTARU["20171023/20181022"]
fedfunds <- data.frame(symbol = "_FEDFUNDS_",
                       date = index(DFEDTARU),
                       price = unname(DFEDTARU),
                       stringsAsFactors = FALSE)
}

}
\keyword{datasets}
