% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal.backtest.R
\docType{package}
\name{tidal.backtest}
\alias{tidal.backtest}
\alias{tidal.backtest-package}
\title{tidal.backtest: Function(s) for Fund Level Backtesting}
\description{
Generate daily NAV backtest results quickly and easily
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
