% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidalbacktest.R
\docType{package}
\name{tidalbacktest}
\alias{tidalbacktest}
\alias{tidalbacktest-package}
\title{tidalbacktest: Function(s) for Fund Level Backtesting}
\description{
Generate daily NAV backtest results quickly and easily.
}
\author{
\strong{Maintainer}: Jason Muhlenkamp \email{jason.muhlenkamp@gmail.com}

}
