% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_weights.R
\name{trade_weights}
\alias{trade_weights}
\title{Trade a Set of Daily Weights}
\usage{
trade_weights(df_prices, df_weights)
}
\arguments{
\item{df_prices}{a data.frame containing the columns: symbol, date, price}

\item{df_weights}{a data.frame containing the columns: symbol, date, w}
}
\description{
Takes a set of daily prices and daily weights
and returns a daily NAV
}
