% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_weights_execute.R
\name{trade_weights_execute}
\alias{trade_weights_execute}
\title{Unexported Subfunction(s) of trade_weights}
\usage{
trade_weights_execute(list_data)
}
\arguments{
\item{list_data}{list returned from trade_weights_initialize()}
}
\description{
Takes the list returned from trade_weights_initialize()
and executes the daily trades.
}
