% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_weights_subfunctions.R
\name{trade_weights_initialize}
\alias{trade_weights_initialize}
\title{Unexported Subfunction of trade_weights}
\usage{
trade_weights_initialize(df_prices, df_weights)
}
\arguments{
\item{df_prices}{a data.frame containing the columns: symbol, date, price}

\item{df_weights}{a data.frame containing the columns: symbol, date, w}
}
\description{
Takes a set of daily prices and daily weights
and returns a list of initialized data ready to be traded
}
