% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_weights_initialize.R
\name{trade_weights_initialize_rebaldates}
\alias{trade_weights_initialize_rebaldates}
\title{Unexported Subfunction(s) of trade_weights}
\usage{
trade_weights_initialize_rebaldates(dates)
}
\arguments{
\item{dates}{a vector of dates}
}
\description{
Returns a date vector indicating which dates trade_weights_execute()
should performance a rebalance.
}
