% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_weights_subfunctions.R
\name{trade_weights_update_nav}
\alias{trade_weights_update_nav}
\title{Unexported Subfunction(s) of trade_weights}
\usage{
trade_weights_update_nav(dt_fund, dt, idate)
}
\arguments{
\item{dt_fund}{data.table of fund-level information}

\item{dt}{data.table of holdings-level information}

\item{idate}{date to update}
}
\description{
Computes nav values for a given date share shares and prices
at various times (value, trade, close) within \code{dt}.
}
