% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_df_columns}
\alias{validate_df_columns}
\title{Unexported Utility Functions}
\usage{
validate_df_columns(df, columns)
}
\arguments{
\item{df}{a data.frame to validate}

\item{columns}{a character vector of columns that must be present in the data.frame}
}
\description{
Utility functions that do behind the scences work
}
