% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidalbacktest.R
\docType{package}
\name{tidalbacktest}
\alias{tidalbacktest}
\alias{tidalbacktest-package}
\title{Package for Backtesting}
\description{
Contains \code{\link{backtest}} function. \cr\cr
}
\author{
\strong{Maintainer}: Jason Muhlenkamp \email{jason.muhlenkamp@gmail.com}

}
