% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_weights.R
\name{trade_weights}
\alias{trade_weights}
\title{Trade a Set of Daily Weights}
\usage{
trade_weights(df_prices, df_weights, rebal_ith = 1,
  rebal_group_by = "date")
}
\arguments{
\item{df_prices}{a data.frame containing the columns: symbol, date, price}

\item{df_weights}{a data.frame containing the columns: symbol, date, w}

\item{rebal_ith}{Optional Integer to determine which observation of each by group to return.
For Example ith_date = 2, week will return the second day of
each week.  Also, ith_date = 99, month will return last day of
each month.  Default is 1.}

\item{rebal_group_by}{Optional Character indicating by group to apply to ith_date.
Acceptable values are: date, week, month, qtr.  Default is date
which performs a daily rebalance.}
}
\description{
Takes a set of daily prices and daily weights
and returns a daily NAV
}
\examples{
\dontrun{
library(quantmod)

# Create a data.frame of adjusted close prices
getSymbolsData <- new.env()
getSymbols(c('XLB', 'XLE', 'XLF'),
           env = getSymbolsData,
           from = as.Date("2018-09-03"),
           to = as.Date("2018-09-29"))
df_close <- dplyr::bind_rows(
    lapply(getSymbolsData, function(x)
        data.frame(date = index(x), price = unname(Ad(x)))), .id = 'symbol')

# Create a data.frame of weights
df_weights <- df_close
df_weights$w <- 0.32
df_weights$price <- NULL

# Trade the weights
df_nav <- trade_weights(df_close, df_weights)
}
}
